fix: error showing in the component
