+++ packages/cfd/src/Containers/cfd-password-modal/modal-elements/password-modal-message.tsx
@@ -1,8 +1,16 @@
 import React from 'react';
-import { CFD_PLATFORMS, getCFDPlatformLabel, Jurisdiction } from '@deriv/shared';
-import { getFormattedJurisdictionCode } from '../../../Stores/Modules/CFD/Helpers/cfd-config';
+import { CFD_PLATFORMS, getCFDPlatformLabel, Jurisdiction, getCategoryText } from '@deriv/shared';
+import {
+    getDerivezCompanies,
+    getDxCompanies,
+    getFormattedJurisdictionCode,
+    getMtCompanies,
+    TDerivezCompanies,
+    TDxCompanies,
+    TMtCompanies,
+} from '../../../Stores/Modules/CFD/Helpers/cfd-config';
 import { Localize } from '@deriv/translations';
-import { getCategoryLabel, getTypeLabel } from 'Constants/cfd-password-modal-content/cfd-password-modal-content';
+import { getTypeLabel } from '../../../Constants/cfd-password-modal-content/cfd-password-modal-content';
 
 type TReviewMsgForMT5 = {
     is_selected_mt5_verified: boolean;
@@ -52,7 +60,31 @@ const PasswordModalMessage = ({
 }: TPasswordModalMessage) => {
     if (!category && !type) return null;
 
-    const category_label = getCategoryLabel(category);
+    const category_label = getCategoryText(category);
+
+    let type_label = '';
+    switch (platform) {
+        case CFD_PLATFORMS.MT5:
+            type_label =
+                getMtCompanies(show_eu_related_content)[category as keyof TMtCompanies][
+                    type as keyof TMtCompanies['demo' | 'real']
+                ].short_title;
+            break;
+        case CFD_PLATFORMS.DXTRADE:
+            type_label =
+                getDxCompanies()[category as keyof TDxCompanies][type as keyof TDxCompanies['demo' | 'real']]
+                    .short_title;
+            break;
+        case CFD_PLATFORMS.DERIVEZ:
+            type_label =
+                getDerivezCompanies()[category as keyof TDerivezCompanies][
+                    type as keyof TDerivezCompanies['demo' | 'real']
+                ].short_title;
+            break;
+        default:
+            type_label = '';
+            break;
+    }
 
     const jurisdiction_label =
         jurisdiction_selected_shortcode && getFormattedJurisdictionCode(jurisdiction_selected_shortcode);
@@ -77,7 +109,7 @@ const PasswordModalMessage = ({
                 <Localize
                     i18n_default_text='Congratulations, you have successfully created your {{category}} <0>{{platform}}</0> <1>{{type}} {{jurisdiction_selected_shortcode}}</1> account. '
                     values={{
-                        type: getTypeLabel(type, platform, show_eu_related_content, category),
+                        type: type_label,
                         platform: platform === CFD_PLATFORMS.MT5 ? mt5_platform_label : getCFDPlatformLabel(platform),
                         category: category_label,
                         jurisdiction_selected_shortcode:
@@ -112,7 +144,7 @@ const PasswordModalMessage = ({
                 <Localize
                     i18n_default_text='Congratulations, you have successfully created your {{category}} <0>{{platform}}</0> <1>{{type}}</1> account. '
                     values={{
-                        type: getTypeLabel(type, platform, show_eu_related_content, category),
+                        type: type_label,
                         platform: getCFDPlatformLabel(platform),
                         category: category_label,
                     }}
