+++ packages/components/src/components/icon/icons.js
@@ -6,7 +6,6 @@ import './appstore/ic-appstore-add.svg';
 import './appstore/ic-appstore-app-card-default-background-small.svg';
 import './appstore/ic-appstore-app-card-default-background.svg';
 import './appstore/ic-appstore-cfds-trade-type.svg';
-import './appstore/ic-appstore-cfds.svg';
 import './appstore/ic-appstore-check.svg';
 import './appstore/ic-appstore-checked-red-bg.svg';
 import './appstore/ic-appstore-checked-transparent.svg';
@@ -15,12 +14,8 @@ import './appstore/ic-appstore-circle.svg';
 import './appstore/ic-appstore-close-dark.svg';
 import './appstore/ic-appstore-close-light.svg';
 import './appstore/ic-appstore-deriv-logo.svg';
-import './appstore/ic-appstore-deriv-x.svg';
-import './appstore/ic-appstore-derived.svg';
-import './appstore/ic-appstore-derivez.svg';
 import './appstore/ic-appstore-ellipse.svg';
 import './appstore/ic-appstore-error.svg';
-import './appstore/ic-appstore-financial.svg';
 import './appstore/ic-appstore-get-more-accounts.svg';
 import './appstore/ic-appstore-home-dark.svg';
 import './appstore/ic-appstore-home.svg';
@@ -32,7 +27,6 @@ import './appstore/ic-appstore-multipliers-trade-type.svg';
 import './appstore/ic-appstore-option-trade-type.svg';
 import './appstore/ic-appstore-skeleton-card-border.svg';
 import './appstore/ic-appstore-success.svg';
-import './appstore/ic-appstore-swap-free.svg';
 import './appstore/ic-appstore-tick-white.svg';
 import './appstore/ic-appstore-tick.svg';
 import './appstore/ic-appstore-traders-hub-home.svg';
