+++ packages/components/stories/icon/icons.js
@@ -305,6 +305,7 @@ export const icons =
         'IcBotBuilderArrow',
         'IcBotBuilderTabIcon',
         'IcBotBuilder',
+        'IcBotStop',
         'IcBox',
         'IcButtonBack',
         'IcCalendarDatefrom',
